Created script to run DOOM III from command-line.
This is just a simple script for me to have an easy way to do the
whole CMake generate and build step, and finally launching the game
with a single command.

Added an ignore the the Git ignore for the build folder as these
files are the CMake and build artifacts that can be regenerated as
needed. There is no need to add these to the repository.

Signed-off-by: Wynand Pieterse <wpieterse2825@gmail.com>